+++ Popup/Popup.js
@@ -27,11 +27,19 @@ import PopupState from './PopupState';
 
 import componentCss from './Popup.module.less';
 
-
 const TransitionContainer = SpotlightContainerDecorator(
 	{enterTo: 'default-element', preserveId: true},
 	Transition
 );
+
+const transitionDirection = {
+	bottom: 'down',
+	center: 'down',
+	fullscreen: 'down',
+	left: 'left',
+	right: 'right',
+	top: 'up'
+};
 
 /**
  * The base popup component.
@@ -42,8 +50,28 @@ const TransitionContainer = SpotlightContainerDecorator(
  * @public
  */
 const PopupBase = kind({
-	name: 'Popup',
+	name: 'PopupBase',
+
 	propTypes: /** @lends agate/Popup.PopupBase.prototype */ {
+		/**
+		 * Set the priority with which screen reader should treat updates to live regions for the Popup.
+		 *
+		 * @type {String|Object}
+		 * @public
+		 */
+		'aria-live': PropTypes.oneOfType([
+			PropTypes.string,
+			PropTypes.object
+		]),
+
+		/**
+		 * Buttons to be included within the Popup component.
+		 *
+		 * Typically, these buttons would be used to close or take action on the dialog.
+		 *
+		 * @type {Element|Element[]}
+		 * @public
+		 */
 		buttons: PropTypes.oneOfType([
 			PropTypes.element,
 			PropTypes.arrayOf(PropTypes.element)
@@ -58,64 +86,230 @@ const PopupBase = kind({
 		 */
 		centered: PropTypes.bool,
 
+		/**
+		 * When true, popup displays a close button.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
 		closeButton: PropTypes.bool,
+
+		/**
+		 * Customizes the component by mapping the supplied collection of CSS class names to the
+		 * corresponding internal elements and states of this component.
+		 *
+		 * The following classes are supported:
+		 *
+		 * * `popup` - The root class name
+		 * * `body` - Applied to the body content container
+		 * * `popupTransitionContainer` - Applied to the Popup's outermost container. Sizing can be
+		 *                                applied here for percentage-of-screen values.
+		 * * `top` - Applied when the `position` is 'top'
+		 * * `right` - Applied when the `position` is 'right'
+		 * * `bottom` - Applied when the `position` is 'bottom'
+		 * * `left` - Applied when the `position` is 'left'
+		 *
+		 * @type {Object}
+		 * @public
+		 */
 		css: PropTypes.object,
+
+		/**
+		 * Controls how long the transition should take.
+		 * Supported preset durations are: `'short'` (250ms), `'medium'` (500ms), and `'long'` (1s).
+		 * `'short'` (250ms) is default when no others are specified.
+		 *
+		 * @type {String}
+		 * @default 'short'
+		 * @public
+		 */
+		duration: PropTypes.string,
+
+		/**
+		 * Support accessibility options.
+		 *
+		 * If true, the aria-live and role in Popup are `null`.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @private
+		 */
+		noAlertRole: PropTypes.bool,
+
+		/**
+		 * Disables transition animation.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
 		noAnimation: PropTypes.bool,
+
+		/**
+		 * Called when the popup is closed.
+		 *
+		 * @type {Function}
+		 * @public
+		 */
 		onClose: PropTypes.func,
+
+		/**
+		 * Called after the popup's "hide" transition finishes.
+		 *
+		 * @type {Function}
+		 * @public
+		 */
 		onHide: PropTypes.func,
+
+		/**
+		 * Called after the popup's "show" transition finishes.
+		 *
+		 * @type {Function}
+		 * @public
+		 */
+		onShow: PropTypes.func,
+
+		/**
+		 * Controls the visibility of the Popup.
+		 *
+		 * By default, the Popup and its contents are not rendered until `open`.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
 		open: PropTypes.bool,
 
 		/**
 		 * Sets the position of the popup on the screen.
 		 *
-		 * @type {('center'|'top')}
+		 * @type {('bottom'|'center'|'fullscreen'|'left'|'right'|'top')}
 		 * @default 'center'
 		 * @public
 		 */
-		position: PropTypes.oneOf(['center', 'top']),
+		position: PropTypes.oneOf(['bottom', 'center', 'fullscreen', 'left', 'right', 'top']),
 
+		/**
+		 * The ARIA role for the Popup.
+		 *
+		 * @type {String|Object}
+		 * @public
+		 */
+		role: PropTypes.oneOfType([
+			PropTypes.string,
+			PropTypes.object
+		]),
+
+		/**
+		 * The current skin.
+		 *
+		 * @type {String}
+		 * @private
+		 */
 		skin: PropTypes.string,
-		title: PropTypes.string
+
+		/**
+		 * The container id for {@link spotlight/Spotlight}.
+		 *
+		 * @type {String}
+		 * @default null
+		 * @public
+		 */
+		spotlightId: PropTypes.string,
+
+		/**
+		 * Restricts or prioritizes navigation when focus attempts to leave the popup.
+		 *
+		 * It can be either `'none'`, `'self-first'`, or `'self-only'`.
+		 *
+		 * Note: The ready-to-use [Popup]{@link agate/Popup.Popup} component only supports
+		 * `'self-first'` and `'self-only'`.
+		 *
+		 * @type {('none'|'self-first'|'self-only')}
+		 * @default 'self-only'
+		 * @public
+		 */
+		spotlightRestrict: PropTypes.oneOf(['none', 'self-first', 'self-only']),
+
+		/**
+		 * The primary text of the popup.
+		 *
+		 * @type {String}
+		 * @public
+		 */
+		title: PropTypes.string,
+
+		/**
+		 * The type of transition to affect the content.
+		 *
+		 * Supported types are: `'slide'` and `'fade'`.
+		 *
+		 * @type {String}
+		 * @public
+		 */
+		type: PropTypes.oneOf(['fade', 'slide'])
 	},
 	defaultProps: {
 		centered: false,
 		closeButton: false,
+		duration: 'short',
+		noAlertRole: false,
 		noAnimation: false,
 		open: false,
-		position: 'center'
+		position: 'center',
+		spotlightRestrict: 'self-only'
 	},
 	styles: {
 		css: componentCss,
-		className: 'popup'
+		className: 'popup',
+		publicClassNames: ['popup', 'body', 'popupTransitionContainer', 'top', 'right', 'bottom', 'left']
 	},
 	computed: {
-		className: ({centered, closeButton, position, styler, title}) => styler.append({top: position === 'top', withCloseButton: closeButton, withTitle: title, centered}),
-		transitionType : ({position}) => position === 'center' ? 'fade' : 'slide',
-		direction : ({position}) => position === 'center' ? 'down' : 'up'
+		// When passing `aria-live` prop to the Popup, the prop should work first.
+		// If `noAlertRole` is true, alert role and aria-live will be removed. Contents of the popup won't be read automatically when opened.
+		// Otherwise, `aria-live` will be usually `off`.
+		'aria-live': ({'aria-live': live, noAlertRole}) => ((typeof live !== 'undefined') ? live : (!noAlertRole && 'off' || null)),
+		className: ({centered, closeButton, position, styler, title}) => styler.append(position, {withCloseButton: closeButton, withTitle: title, centered}),
+		direction: ({position}) => transitionDirection[position],
+		// When passing `role` prop to the Popup, the prop should work first.
+		// If `noAlertRole` is true, alert role and aria-live will be removed. Contents of the popup won't be read automatically when opened.
+		// Otherwise, `role` will be usually `alert`.
+		role: ({noAlertRole, role}) => ((typeof role !== 'undefined') ? role : (!noAlertRole && 'alert' || null)),
+		transitionContainerClassName: ({css, position, styler}) => styler.join(css.popupTransitionContainer, position),
+		transitionType: ({position, type}) => {
+			const setTransitionType = position === 'center' ? 'fade' : 'slide';
+
+			return type ? type : setTransitionType;
+		}
 	},
-	render: ({buttons, children, closeButton, css, direction, noAnimation, onClose, onHide, open, skin, title, transitionType, ...rest}) => {
+	render: ({buttons, children, closeButton, css, direction, duration, noAnimation, onClose, onHide, onShow, open, skin, spotlightId, spotlightRestrict, title, transitionContainerClassName, transitionType, ...rest}) => {
 		const wideLayout = (skin === 'carbon');
 		delete rest.centered;
+		delete rest.noAlertRole;
 
 		return (
 			<TransitionContainer
-				noAnimation={noAnimation}
-				visible={open}
-				direction={direction}
-				duration="short"
-				type={transitionType}
-				className={css.popupTransitionContainer}
-				onHide={onHide}
+				className={transitionContainerClassName}
 				css={css}
+				direction={direction}
+				duration={duration}
+				noAnimation={noAnimation}
+				onHide={onHide}
+				onShow={onShow}
+				spotlightId={spotlightId}
+				spotlightRestrict={spotlightRestrict}
+				type={transitionType}
+				visible={open}
 			>
 				<div
 					role="alert"
 					{...rest}
 				>
 					{closeButton ? <Button
+						className={componentCss.closeButton}
 						icon="closex"
 						onTap={onClose}
-						className={componentCss.closeButton}
 						size="small"
 					/> : null}
 					{title ? <Heading className={css.title}>{title}</Heading> : null}